see RN2.4.2 (20200930)

+++ aTalk/src/main/java/org/atalk/service/version/VersionService.java
@@ -24,7 +24,7 @@ public interface VersionService
 	 */
 	Version getCurrentVersion();
 
-	int getCurrentVersionCode();
+	long getCurrentVersionCode();
 
 	String getCurrentVersionName();
